+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log
 
 - Fix an issue that “Open Scripts Folder” command in the script menu did not open the scripts folder for CotEditor but just the general Application Scripts folder.
 - Fix an issue in the find result table that the found string column truncated the text even in the middle of a character that consists of multiple Unicode characters.
+- Fix an issue where line number view did not update under some specific conditions even when the line height setting is changed.
 - Fix the progress message when highlighting a document with a multiple replacement definition.